fixed wrong casts with hashmap and pointers
When inserting pointers (e.g. to variables or strings), we should
not first cast to (size_t), since this might have a smaller size
than a pointer. On Linux-Systems this is (so far) not the case, so
that so far no problem occured.

+++ src/scip/heur_crossover.c
@@ -888,7 +888,7 @@ SCIP_DECL_HEUREXEC(heurExecCrossover)
    SCIP_CALL( SCIPallocBufferArray(scip, &selection, nusedsols) );
 
    for( i = 0; i < nvars; i++ )
-     subvars[i] = (SCIP_VAR*) (size_t) SCIPhashmapGetImage(varmapfw, vars[i]);
+     subvars[i] = (SCIP_VAR*) SCIPhashmapGetImage(varmapfw, vars[i]);
 
    /* free hash map */
    SCIPhashmapFree(&varmapfw);
